+++ example/reflect/cmd/client.go
