+++ example/reflect/cmd/pkg-vars.go
