+++ example/reflect/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/michaelquigley/pfxlog"
-	"github.com/openziti/sdk-golang/example/echo/cmd"
+	"github.com/openziti/sdk-golang/example/reflect/cmd"
 	"github.com/openziti/sdk-golang/ziti/config"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
